refactor to make common in computational and dynamic
